Create dockerfile
A dockerfile is used to build an image of the application that can
then be launched in a container.

It will begin from a python 3.6 image, and add the contents in the
current directory to /app in the container. The working directory
will then be changed to /app, install the required python modules
and then expose port 8000 in the container. The primary process of
the container is then set.

The commands run once during the creation, except for CMD, which
runs when the container is created.